requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
PyJWT==2.8.0
six==1.16.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-filter==23.5
django-allauth==0.57.0
requests-oauthlib==1.3.1
